+++ frontend/src/_test_/Edit_organisations.test.js
@@ -1,13 +1,13 @@
 import React from 'react'
 import { shallow, configure } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import Search from '../../Components/Organisation/Search';
+import EditOrganisation from '../Components/Organisation/EditOrganisation';
 
 configure({ adapter: new Adapter() });
 
-describe('Search Component', () => {
+describe('Edit Organization Component', () => {
     it('Should render without crashing', () => {
-        const wrapper = shallow(<Search />).length
+        const wrapper = shallow(<EditOrganisation />).length
         expect(wrapper).toEqual(1)
     })
 })
