+++ frontend/src/_test_/Organisation_form.test.js
@@ -3,17 +3,17 @@ import { shallow, configure } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
 import { Provider } from 'react-redux';
 import { createStore } from 'redux';
-import reducerRoot from '../../reducers';
-import AddOrganisation from '../../Components/Organisation/AddOrganisation';
+import reducerRoot from '../reducers';
+import OrganizationForm from '../Components/Organisation/OrganisationForm';
 
 configure({ adapter: new Adapter() });
 
 const store = createStore(reducerRoot);
 
-describe('Add Organization Component', () => {
+describe('OrganizationForm Component', () => {
     it('Should render without crashing', () => {
         const wrapper = 
-        shallow(<Provider store={store}><AddOrganisation /></Provider>).length
+        shallow(<Provider store={store}><OrganizationForm /></Provider>).length
         expect(wrapper).toEqual(1)
     })
 })
