+++ frontend/src/_test_/Single_organisation.test.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { shallow, configure } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import SingleOrganization from '../../Components/Organisation/SingleOrganisation';
+import SingleOrganization from '../Components/Organisation/SingleOrganisation';
 
 configure({ adapter: new Adapter() });
